Renamed data type validations
These validations are more accurately named with a prefix of DataType
instead of Validate. A prefix was needed as some of the data types are
reserved words, so for consistency all data types were given the prefix.

They are being renamed in the first place because the abstract class is
needs to be further abstracted.
Affects #15
Task #15 - Add support for default values