Feature 15 support default fields (#22)
* Renamed data type validations

These validations are more accurately named with a prefix of DataType
instead of Validate. A prefix was needed as some of the data types are
reserved words, so for consistency all data types were given the prefix.

They are being renamed in the first place because the abstract class is
needs to be further abstracted.
Affects #15
Task #15 - Add support for default values

* Refactored abstract validate and added abstract data type

For schema validations the majority of the logic in abstract validate
was needed. So code was refactored to better represent the name spaces.

Task #15

* Added model to Abstract Validate

Model was added to abstract validate to allow other validations the 
opportunity to interact with the core model. This gives the flexibility
to execute events, access / use properties and direct access to the data

Task #15

* Added Params to Abstract Validate

Task #15

* Increased test depth against AbstractValidate::getParams

Task #15

* Added a pre execute check to AbstractValidate

This is a template that can be overridden to get access to mroe control
on what executes and doesn't execute. Control logic was poisoning the
business logic of the validate function.

Task #15

* Added default value validation

Values will now be added to the create event. Will ignore all other
events.

Task #15

* Corrected layout to meet PSR-2 standards


Task #15

* Updated documentation to reflect new Default value validation

Task #15